backend/requirements.txt
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

langchain==0.3.23
langchain-groq==0.3.2
langchain-google-genai==2.1.3

langgraph==0.3.31
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8

fastapi==0.115.9
uvicorn==0.31.0
pydantic==2.10.6
asyncio==3.4.3

weaviate==0.1.2
weaviate-client==4.13.2

openpyxl==3.1.4
pandas==2.2.3
sentence-transformers==4.1.0
SQLAlchemy==2.0.35

mysql-connector-python==9.3.0
PyMySQL==1.1.1
psycopg2-binary==2.9.10
pyodbc==5.2.0

python-dotenv==1.0.1